refactor: Add path alias
