+++ jest.config.js
@@ -3,8 +3,14 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   moduleNameMapper: {
-    '@assets/(.*)': '<rootDir>/assets/$1',
-    '@src/(.*)': '<rootDir>/src/$1',
+    '@assets/(.*)': '<rootDir>/src/assets/$1',
+    '@defines/(.*)': '<rootDir>/src/defines/$1',
+    '@backend/(.*)': '<rootDir>/src/backend/$1',
+    '@frontend/(.*)': '<rootDir>/src/frontend/$1',
+    '@pages/(.*)': '<rootDir>/src/pages/$1',
+    '@types': '<rootDir>/src/types',
+    '@utils/(.*)': '<rootDir>/src/utils/$1',
+    '$src/(.*)': '<rootDir>/src/$1',
   },
   moduleDirectories: ['node_modules'],
   testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/.next/'],
