+++ src/pages/_app.tsx
@@ -5,10 +5,10 @@ import { DefaultSeo } from 'next-seo';
 import useNProgress from 'next-use-nprogress';
 import Script from 'next/script';
 
-import { CommonLayout } from '@src/frontend/components/layout';
-import { Modal, Notification } from '@src/frontend/components/ui';
-import { useModal } from '@src/frontend/hooks/use-modal';
-import { useNoti } from '@src/frontend/hooks/use-noti';
+import { CommonLayout } from '@frontend/components/layout';
+import { Modal, Notification } from '@frontend/components/ui';
+import { useModal } from '@frontend/hooks/use-modal';
+import { useNoti } from '@frontend/hooks/use-noti';
 
 import type { AppProps } from 'next/app';
