+++ src/pages/api/version/index.test.ts
@@ -1,6 +1,6 @@
 import { StatusCodes } from 'http-status-codes';
 
-import TH from '@src/backend/test-helper';
+import TH from '@backend/test-helper';
 
 import apiHandler, { API_VERSION } from '.';
