+++ src/pages/loading.tsx
@@ -1,4 +1,4 @@
-import { Loading } from '@src/frontend/components/core';
+import { Loading } from '@frontend/components/core';
 
 export default function LoadingPage() {
   return <Loading />;
