+++ src/utils/mongodb/index.ts
@@ -1,6 +1,6 @@
 import { MongoClient, type MongoClientOptions } from 'mongodb';
 
-import { isDev, isTest, MONGODB_URI } from '@src/utils/env';
+import { isDev, isTest, MONGODB_URI } from '@utils/env';
 
 const options: MongoClientOptions = {
   ignoreUndefined: true,
